Crate rosidl_runtime_rs

Bindings to rosidl_runtime_c and related functionality for messages.

Macros

seq

Creates a sequence, similar to the vec! macro.

Structs

BoundedSequence

A bounded sequence.

BoundedString

A zero-terminated UTF-8 string with a length limit.

BoundedWString

A zero-terminated UTF-16 string with a length limit.

Sequence

An unbounded sequence.

SequenceExceedsBoundsError

Error type for BoundedSequence::try_new().

String

A zero-terminated UTF-8 string.

StringExceedsBoundsError

Error type for BoundedString::try_from() and BoundedWString::try_from().

WString

A zero-terminated UTF-16 string.

Traits

Message

Trait for types that can be used in a rclrs::Subscription and a rclrs::Publisher.

RmwMessage

Trait for RMW-native messages.

SequenceAlloc

Internal trait that connects a particular Sequence<T> instance to generated C functions that allocate and deallocate memory.

Service

Trait for services.