Trait rosidl_runtime_rs::RmwMessage

pub trait RmwMessage: Clone + Debug + Default + Send + Sync + Message {
    const TYPE_NAME: &'static str;

    // Required method
    fn get_type_support() -> *const c_void;
}

Trait for RMW-native messages.

See the documentation for the Message trait, which is the trait that should generally be used by user code.

User code never needs to call this trait’s method, much less implement this trait.

Required Associated Constants

const TYPE_NAME: &'static str

A string representation of this message’s type, e.g. “geometry_msgs/msg/Twist”

Required Methods

fn get_type_support() -> *const c_void

Get a pointer to the correct rosidl_message_type_support_t structure.

Object Safety

This trait is not object safe.