Trait rosidl_runtime_rs::SequenceAlloc

pub trait SequenceAlloc: Sized {
    // Required methods
    fn sequence_init(seq: &mut Sequence<Self>, size: usize) -> bool;
    fn sequence_fini(seq: &mut Sequence<Self>);
    fn sequence_copy(
        in_seq: &Sequence<Self>,
        out_seq: &mut Sequence<Self>
    ) -> bool;
}

Internal trait that connects a particular Sequence<T> instance to generated C functions that allocate and deallocate memory.

User code never needs to call these trait methods, much less implement this trait.

Required Methods

fn sequence_init(seq: &mut Sequence<Self>, size: usize) -> bool

Wraps the corresponding init function generated by rosidl_generator_c.

fn sequence_fini(seq: &mut Sequence<Self>)

Wraps the corresponding fini function generated by rosidl_generator_c.

fn sequence_copy(in_seq: &Sequence<Self>, out_seq: &mut Sequence<Self>) -> bool

Wraps the corresponding copy function generated by rosidl_generator_c.

Object Safety

This trait is not object safe.